+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   New component in React : `Link`
+-   New component in React : `WebBookmark`
 
 ## [0.21.8][] - 2020-02-19
 
@@ -22,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   Added `focusPoint` prop to `Thumbnail` component
+
 
 ## [0.21.6][] - 2020-02-13